Import from revision 38398 from trunk: When temporarily stroing the 'struct offset', make sure to use the right type (it used to be 'int' and has been changed to 'long' but a few instances escaped the proper updates). This fixes: <http://savannah.cern.ch/bugs/?75874>
git-svn-id: http://root.cern.ch/svn/root/branches/v5-28-00-patches@38399 27541ba8-7e3a-0410-8455-c3a389f83636

+++ cint/cint/src/ifunc.cxx
@@ -4870,7 +4870,7 @@ int G__interpret_func(G__value* result7, const char* funcname, G__param* libp, i
    struct G__ifunc_table_internal* store_p_ifunc = p_ifunc;
    int specialflag = 0;
    G__value* store_p_tempobject = 0;
-   int store_memberfunc_struct_offset;
+   long store_memberfunc_struct_offset;
    int store_memberfunc_tagnum;
 #ifndef G__OLDIMPLEMENTATION2038
    G_local.enclosing_scope = 0;
